chore(config): Generate JSON schema with typescript-json-schema
