+++ packages/config/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./config";
+export * from "./types";
 export * from "./validate";
