+++ packages/config/src/types.ts
@@ -1,0 +1,8 @@
+export interface EnvironmentConfig {
+  require?: ReadonlyArray<string>;
+  components?: ReadonlyArray<string>;
+}
+
+export interface Config extends EnvironmentConfig {
+  environments?: { [key: string]: EnvironmentConfig };
+}
